requirements.txt
django==2.2.0
django-filter==0.2.4
django-bootstrap-form==0.2.4